% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmdC.homoVAR.fn.R
\name{ssmdC.homoVAR.fn}
\alias{ssmdC.homoVAR.fn}
\title{calculate the critical value of SSMD}
\usage{
ssmdC.homoVAR.fn(n1, n2, Alpha = 0.05, Beta = 3)
}
\arguments{
\item{n1}{a value or a vector for the first group (i.e., the reference group)}

\item{n2}{a vector for the second group}

\item{Alpha}{specify the significant level}

\item{Beta}{the population value of SSMD}
}
\value{
a value
}
\description{
A function to calculate the critical value of SSMD
}
\examples{
ssmdC.homoVAR.fn(304, 16, Alpha=0.05, Beta=3)
}
\author{
Xiaohua Douglas Zhang   11/2019
}
