% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmdEst.homoVAR.fn.R
\name{ssmdEst.homoVAR.fn}
\alias{ssmdEst.homoVAR.fn}
\title{calculate UMVUE estimate of SSMD between two groups under homoscedasticity}
\usage{
ssmdEst.homoVAR.fn(x, y, pXtrim = 0, approx = FALSE, method = "UMVUE")
}
\arguments{
\item{x}{the values of response in the first group (i.e., the reference group)}

\item{y}{the values of response in the second group}

\item{approx}{approx=TRUE, indicating the use of approximation for the coefficient in the UMVUE estimate of SSMD}

\item{method}{specify the method for estimating SSMD. It must be either 'UMVUE', 'MLE' or'MM'.}
}
\value{
a vector
}
\description{
A function to calculate UMVUE estimate of SSMD between two groups under homoscedasticity
}
\examples{
x = rnorm(320, mean=0, sd=1)
  y = rnorm(8, mean=3*sqrt(2), sd=1)
  ssmdEst.homoVAR.fn( x, y, pXtrim=0.05, approx=F, method="UMVUE")
}
\author{
Xiaohua Douglas Zhang originally in 2007, revised in 2020
}
