% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmdEst.homoVAR.frame.fn.R
\name{ssmdEst.homoVAR.frame.fn}
\alias{ssmdEst.homoVAR.frame.fn}
\title{calculate UMVUE estimate of SSMD between a positive control and
  a negative reference in a data frame for an HTS study}
\usage{
ssmdEst.homoVAR.frame.fn(
  dataIn.df,
  negREF = "Sample",
  positiveCTRL = "positive control",
  pREFtrim = 0,
  approx = FALSE,
  method = "UMVUE"
)
}
\arguments{
\item{dataIn.df}{a data frame with plate ID (or name), well usage and intensity}

\item{negREF}{name of wells to be used as a negative reference}

\item{positiveCTRL}{name of wells to be used as a positive control}

\item{pREFtrim}{the total portion of data to be trimmed in the negative reference}

\item{approx}{approx=TRUE, indicating the use of approximation for the coefficient
in the UMVUE estimate of SSMD}

\item{method}{specify the method for estimating SSMD. It must be
either 'UMVUE', 'MLE' or'MM'. "}
}
\value{
a matrix containing estimated SSMD value and variance for each plate
}
\description{
A function to calculate UMVUE estimate of SSMD between a positive control and
  a negative reference in a data frame for an HTS study
}
\examples{
 data("data.CVB3CRISPR", package="qcSSMDhomo")
  data.df = dataCVB3.df
  ssmdEst.homoVAR.frame.fn(dataIn.df=data.df, negREF="Sample", positiveCTRL="Positive Control",
                         pREFtrim=0.05, Alpha=0.05, Beta=3)


}
\author{
Xiaohua Douglas Zhang and Dandan Wang        07/2020
}
