% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmdEstCore.homoVAR.fn.R
\name{ssmdEstCore.homoVAR.fn}
\alias{ssmdEstCore.homoVAR.fn}
\title{calculate UMVUE estimate of SSMD}
\usage{
ssmdEstCore.homoVAR.fn(
  m1,
  m2,
  s1,
  s2,
  n1,
  n2,
  approx = FALSE,
  method = "UMVUE"
)
}
\arguments{
\item{m1}{sample mean of values in the first group (i.e., the reference group)}

\item{m2}{sample mean of values in the second group}

\item{s1}{sample standard deviation of values in the first group (i.e., the reference group)}

\item{s2}{sample standard deviation of values in the second group}

\item{n1}{sample size in the first group (i.e., the reference group)}

\item{n2}{sample size in the second group}

\item{approx}{approx=TRUE, indicating the use of approximation for the coefficient in the UMVUE estimate of SSMD}

\item{method:}{specify the method for estimating SSMD. It must be}
}
\value{
a vector
}
\description{
A function to calculate UMVUE estimate of SSMD
}
\examples{
 x = rnorm(304, mean=0, sd=1)
  y = rnorm(8, mean=3*sqrt(2), sd=1)
  ssmdEstCore.homoVAR.fn( mean(x), mean(y), sd(x), sd(y), length(x), length(y), method="UMVUE") 

}
\author{
Xiaohua Douglas Zhang  originally in 2007, revised in 2020
}
