% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zFactor.frame.fn.R
\name{zFactor.frame.fn}
\alias{zFactor.frame.fn}
\title{calculate the plug-in value of z-factor between a positive control and}
\usage{
zFactor.frame.fn(
  dataIn.df,
  negREF = "Sample",
  positiveCTRL = "posCTRL",
  pREFtrim = 0,
  is.homoVAR = TRUE,
  k = 3
)
}
\arguments{
\item{dataIn.df}{a data frame with plate ID (or name), well usage and intensity}

\item{negREF}{name of wells to be used as a negative reference}

\item{positiveCTRL}{name of wells to be used as a positive control}

\item{pREFtrim}{the total portion of data to be trimmed in the negative reference}

\item{is.homoVAR}{indicator for whether to use homoscedasticity or heteroscedasticity}

\item{k:}{the coefficient in z-factor}
}
\value{
a matrix containing estimated plug-in z-factor for each plate
}
\description{
A function to calculate the plug-in value of z-factor between a positive control and
}
\examples{
 data("data.CVB3CRISPR", package="qcSSMDhomo")
  data.df = dataCVB3.df
  zFactor.frame.fn(dataIn.df=data.df, negREF="Sample", positiveCTRL="Positive Control",
                   pREFtrim=0.05, pREFtrim=0, is.homoVAR=TRUE, k=3)

}
\author{
Xiaohua Douglas Zhang and Dandan Wang        07/2020
}
